requirements.txt
# Core ML/Data Science
numpy>=1.26.0           # 1.24.x has no wheels for Python 3.12+
pandas>=2.1.0
tensorflow>=2.15.0      # 2.13 is incompatible with Python 3.12+
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0

# Utilities
pyyaml>=6.0              # Central config loading (config.yaml)
python-dateutil>=2.8.0
joblib>=1.3.0            # Model artifact serialization

# Additional dependencies
openpyxl>=3.1.0          # For Excel file handling (if you save .xlsx)
Pillow>=10.0.0           # For image processing in visualizations
h5py>=3.9.0              # For saving Keras models (.h5 format)

# Optional but recommended
jupyterlab>=4.0.0        # For Jupyter notebooks (if you use them)
ipykernel>=6.25.0        # Jupyter kernel
